Move render function back to render.rs. Update gpu_context type to be non-mutable
